split main ipc handlers into separate files
