+++ noteworthy-electron/src/main/app.ts
@@ -8,15 +8,7 @@ import { EventEmitter } from "events";
 
 // project imports
 import MainWindow from "./windows/MainWindow";
-import {
-	MainIpc_FileHandlers, MainIpc_TagHandlers, MainIpc_DialogHandlers,
-	MainIpc_LifecycleHandlers, MainIpc_ThemeHandlers, MainIpc_ShellHandlers,
-	MainIpcHandlers, MainIpc_OutlineHandlers,
-	MainIpc_MetadataHandlers,
-	MainIpc_NavigationHandlers,
-	MainIpc_CitationHandlers,
-	MainIpcChannelName
-} from "./MainIPC";
+import { MainIpcHandlers, MainIpcChannelName } from "./MainIPC";
 import { FSAL } from "./fsal/fsal";
 import { invokerFor, FunctionPropertyNames } from "@common/ipc";
 import { IDirEntryMeta, IFileMeta, IPossiblyUntitledFile } from "@common/files";
@@ -24,6 +16,16 @@ import { IpcEvents } from "@common/events";
 import { WorkspaceService, WorkspaceEvent } from "./workspace/workspace-service";
 import { ThemeService, ThemeEvent } from "./theme/theme-service";
 import { PluginService } from "./plugins/plugin-service";
+import { MainIpc_LifecycleHandlers } from "./ipc/lifecycle";
+import { MainIpc_FileHandlers } from "./ipc/file";
+import { MainIpc_DialogHandlers } from "./ipc/dialog";
+import { MainIpc_CitationHandlers } from "./ipc/citation";
+import { MainIpc_MetadataHandlers } from "./ipc/metadata";
+import { MainIpc_NavigationHandlers } from "./ipc/navigation";
+import { MainIpc_OutlineHandlers } from "./ipc/outline";
+import { MainIpc_ShellHandlers } from "./ipc/shell";
+import { MainIpc_TagHandlers } from "./ipc/tag";
+import { MainIpc_ThemeHandlers } from "./ipc/theme";
 
 ////////////////////////////////////////////////////////////
